Extract lists outside of functions
Extracted lists outside of functions. Before the extraction the lists were instantiated each time the functions were called. In addition, the functions themselves are called from loops. This means that the lists are needlessly created and destroyed each time the functions are called. Extracting the lists allows them to be created once and used repeatedly.