Make the list instantiation more pythonic
Make the list instantiation more pythonic. This is done by doing two things. The first one is inlining the `'label'` string. The second one is extending the list in-place with desired string objects. This is done by mapping the integers received from `range` into `str` objects. Then the result is extracted into the instantiated list.