Replace try-except with for-loop with list comprehension
Changed loop algorithm to not need `try-except`. Previously, the loop used with a `try-except`. This was done to catch the run reaching an end.

Now it creates a new list with already filtered elements. This is done via list comprehension. This prevents the usage of a `try-except`, and makes the code more maintainable.

An alternative was considered of using a while loop to avoid the try-except. However this would have been measurably slower.

In addition there was minor refractoring done. Specifically the removal of `str(fps)` from within an f-string (which is needless considering the nature of such string).